Add generateImage function to CreatePost page
This commit adds a new function, , to the  page in the client application. The  function sends a POST request to the  endpoint with a prompt from the form, and then sets the  field of the form to the generated image. If no prompt is provided, an error message is displayed.

Note: This commit assumes that the server application is running on http://localhost:8080. If the server is running on a different URL, this code will need to be modified accordingly.